# Set default values when decoding and omit default values when encoding