--- a/bench/vector_tile.proto
+++ b/bench/vector_tile.proto
@@ -1,4 +1,4 @@
-// Protocol Version 1
+syntax = "proto2";
 
 package vector_tile;