Note: This would be a **breaking** change / **major** revision as it drastically affects the way default values are handled.

According to the [protobuf spec](https://developers.google.com/protocol-buffers/docs/proto3#default), when a message is parsed and it does not contain an element the default value for that type is used.

This commit will provide default for each field type when parsing. In addition, similar to other protobuf encoding implementations, when using `proto3` it will omit fields that are equivalent to the default value.

Omitting default values is only safe in `proto3` because the default value can be overriden in `proto2`.