@mourner 2016-06-28
Thanks for the PR! Can you explain the reason for encoding all default values? The default value can be overridden in proto2, but it's a concern of the decoder — there seems to be no need to actually encode all those values (which is costly and increases file sizes significantly).