@kjvalencik 2016-06-28
It could be done in proto2, but it's not necessarily safe. Here's an example. A server and a client are both using the following definition:

```proto
message Thing {
    optional int32 key = 1;
}
```

Our serializer, notices that you are sending `{ key : 0 }`, which is the default, and omits the value. The deserializer fills in the default value. Everything is good.

Later, the message definition is updated in the server.

```proto
message Thing {
    optional int32 key = 1 [default = 5];
}
```

You send, `{ key : 5 }` and the serialzier omits the value. However, the client is still using the old version and fills in what it thinks is the default and the resulting message is `{ key : 0}`. This is safe in `proto3` because you can't override defaults.

It's possible that this is *not* expected to be a backwards compatible / safe change. I couldn't find a reference in the proto spec or not. If there isn't an expectation that changing defaults is safe (just like changing a type isn't safe), then this should be okay.

Also, wasn't sure if my comment was unclear. In `proto2` my PR won't serialize all defaults, just the ones that are explicitly set--which is no change to current functionality. The change I made for `proto3` is still omit fields even if they are explicitly set.