@mourner 2016-06-28
@kjvalencik I think you're expected to update field numbers when you do any breaking change, whether it's changing field type or any of the modifiers like the default value. You can only expect full roundtrip if you use identical proto file, otherwise it's not guaranteed and any changes must be handled explicitly. So I think we shouldn't write default values in the file regardless of whether it's proto3 or proto2.