@mourner 2016-06-28
> Also, wasn't sure if my comment was unclear. In proto2 my PR won't serialize all defaults, just the ones that are explicitly set--which is no change to current functionality.

This still has too much of an impact. On that sample proto, it's a 60% increase! Encoding time also drops by 55%. Since the primary goal of this particular implementation is performance, I'll resist any changes that are not critical but impact performance so much.

```diff
-    t.equal(buf.length, 125023);
+    t.equal(buf.length, 197699);
```