@kjvalencik 2016-06-28
That length is an unfair comparison because decodes the message, which sets all default values, and then re-encodes it. Because of this, the write is explicitly writing all of those default values.

If you start with a plain object and not a deserialized message the size is exactly the same. If you switch to `proto3`, the size is actually smaller.

Decode time is dropping because it is encoding *much* more data. Again, this is because the comparison is apples to oranges. Much more data is being encoded than before.

Maybe I can switch the example to `proto3` to give a better example? Thanks!