@mourner 2016-06-28
@kjvalencik it's apples to oranges if you know what's happening under the hood. But decoding a tile, doing modifications and then reencoding back is a very common use case, and a sudden size increase will be unexpected for many users IMHO. And my point about compatibility above still holds — I think the behavior should be the same for proto2 and proto3.