@kjvalencik 2016-06-28
What is your recommendation? I can think of a few options:

* Enable the encoding check at serialization check for `proto2` as well.
* Make this an opt-in feature at compile time.
* Deserialize into an object with getters that provide the default values.

The first option seems like the best to me, but you bring up an excellent point about the deserialize / write / re-serialize cycle. I need to test that doesn't negatively affect write speed.