@kjvalencik 2016-06-28
I just ran a benchmark for the new serialization (comparing against defaults for omitting) vs. the old and it's about a 2.5% drop in speed. I'm concerned that this could be made worse with defaults that are not trivially compared (e.g., long strings).

Because of that, I'd want to make defaults *only* be provided in `proto3` *or* make this an opt-in feature. Thoughts?