@mourner 2016-06-28
I wouldn't worry about 2.5%, and doubt default strings that are long enough to cause comparison trouble are common enough. My only worry about default values is the `bytes` type — instantiating an empty buffer is an expensive operation in JavaScript. Maybe we should make the default `null`, just like the value for messages, and make a note about this somewhere in the docs.

What do you think about the following approach:

- **proto2**: 
  - decoding: initially set field default (if given) or type default (`null`, `0`, etc.) in the object
  - encoding: if a value is undefined or strictly equal to default (either field or type), do not encode it
- **proto3**
  - decoding: initially set type default
  - encoding: if a value is undefined or strictly equal to type default, do not encode it

So, the same behavior for both proto versions except that `proto3` ignores explicit `default` statements, without any options.

BTW, this change improves decoding speed on the benchmark a little, because V8 doesn't have to change hidden classes for objects as often as it constructs them during decoding — every message has the same set of fields.