@kjvalencik 2016-06-29
@mourner Just saw your note on my other PR about the `null` on `write` checks. I'll remove that from this PR.