@kjvalencik 2016-06-29
@mourner Updated. I broke the PR into separate commits for each substantial change. Let me know if you would like me to squash.

I re-ran the benchmarks 3 times before and after.

#### Before

```
decode vector tile with pbf x 867 ops/sec ±0.86% (89 runs sampled)
encode vector tile with pbf x 590 ops/sec ±1.05% (83 runs sampled)
decode vector tile with pbf x 863 ops/sec ±0.87% (89 runs sampled)
encode vector tile with pbf x 606 ops/sec ±1.16% (84 runs sampled)
decode vector tile with pbf x 866 ops/sec ±0.86% (88 runs sampled)
encode vector tile with pbf x 593 ops/sec ±1.03% (84 runs sampled)
```

#### After

```
decode vector tile with pbf x 891 ops/sec ±1.18% (88 runs sampled)
encode vector tile with pbf x 609 ops/sec ±1.29% (86 runs sampled)
decode vector tile with pbf x 883 ops/sec ±1.19% (89 runs sampled)
encode vector tile with pbf x 609 ops/sec ±1.15% (82 runs sampled)
decode vector tile with pbf x 894 ops/sec ±1.11% (89 runs sampled)
encode vector tile with pbf x 603 ops/sec ±1.34% (84 runs sampled)
```