@mourner 2016-06-29
@kjvalencik this looks great! I'll do a final review tomorrow and merge.